requirements.txt
httpx
python-dotenv
openai
chromadb
chromadb-client
PyPDF2
regex